net/libp2p: Fix tests
