+++ p2p/src/net/libp2p/mod.rs
@@ -16,7 +16,6 @@
 // limitations under the License.
 //
 // Author(s): A. Altonen
-#![allow(unused)]
 use crate::{
     error::{ConversionError, DialError, P2pError, ProtocolError, PublishError},
     message,
@@ -31,30 +30,26 @@ use async_trait::async_trait;
 use itertools::*;
 use libp2p::{
     core::{upgrade, PeerId},
-    gossipsub::{
-        Gossipsub, GossipsubConfigBuilder, MessageAuthenticity, MessageId, ValidationMode,
-    },
-    identify::{Identify, IdentifyConfig, IdentifyInfo},
-    identity,
-    mdns::Mdns,
-    mplex,
+    gossipsub::MessageId,
+    identify::IdentifyInfo,
+    identity, mplex,
     multiaddr::Protocol,
-    noise, ping,
+    noise,
     request_response::*,
-    swarm::{NetworkBehaviour, SwarmBuilder},
+    swarm::SwarmBuilder,
     tcp::TcpConfig,
     Multiaddr, Transport,
 };
 use logging::log;
 use serialization::{Decode, Encode};
-use std::{iter, num::NonZeroU32, sync::Arc, time::Duration};
+use std::sync::Arc;
 use tokio::sync::{mpsc, oneshot};
 use utils::ensure;
 
 mod backend;
 mod constants;
-mod proto;
 mod sync;
+mod tests;
 mod types;
 
 pub mod behaviour;
@@ -117,7 +112,7 @@ where
 ///
 /// Documentation for libp2p-mdns doesn't mention if `peer_addr` includes the PeerId
 /// so if it doesn't, add it. Otherwise just return the address
-fn parse_discovered_addr(peer_id: PeerId, peer_addr: Multiaddr) -> Option<Multiaddr> {
+pub fn parse_discovered_addr(peer_id: PeerId, peer_addr: Multiaddr) -> Option<Multiaddr> {
     let mut components = peer_addr.iter();
 
     if !std::matches!(components.next(), Some(Protocol::Ip4(_) | Protocol::Ip6(_))) {
@@ -172,7 +167,7 @@ where
 }
 
 /// Parse all discovered addresses and group them by PeerId
-fn parse_peers<T>(mut peers: Vec<(PeerId, Multiaddr)>) -> Vec<net::types::AddrInfo<T>>
+pub fn parse_peers<T>(mut peers: Vec<(PeerId, Multiaddr)>) -> Vec<net::types::AddrInfo<T>>
 where
     T: NetworkingService<PeerId = PeerId, Address = Multiaddr>,
 {
@@ -317,7 +312,6 @@ impl NetworkingService for Libp2pService {
     }
 }
 
-// TODO: move services to separate files + unit tests?
 #[async_trait]
 impl<T> ConnectivityService<T> for Libp2pConnectivityHandle<T>
 where
@@ -327,7 +321,6 @@ where
     async fn connect(&mut self, addr: T::Address) -> crate::Result<()> {
         log::debug!("try to establish outbound connection, address {:?}", addr);
 
-        // TODO: add tests for both cases
         let peer_id = match addr.iter().last() {
             Some(Protocol::P2p(hash)) => PeerId::from_multihash(hash).map_err(|_| {
                 P2pError::ConversionError(ConversionError::InvalidAddress(addr.to_string()))
@@ -373,7 +366,6 @@ where
         &self.peer_id
     }
 
-    // TODO: `impl TryInto<ConnectivityEvent> for types::ConnectivityEvent`??
     async fn poll_next(&mut self) -> crate::Result<ConnectivityEvent<T>> {
         match self.conn_rx.recv().await.ok_or(P2pError::ChannelClosed)? {
             types::ConnectivityEvent::ConnectionAccepted { addr, peer_info } => {
@@ -413,7 +405,6 @@ where
     }
 }
 
-// TODO: move services to separate files + unit tests?
 #[async_trait]
 impl<T> PubSubService<T> for Libp2pPubSubHandle<T>
 where
@@ -487,7 +478,6 @@ where
     }
 }
 
-// TODO: move services to separate files + unit tests?
 #[async_trait]
 impl<T> SyncingCodecService<T> for Libp2pSyncHandle<T>
 where
@@ -583,354 +573,5 @@ where
                 error,
             }),
         }
-    }
-}
-
-// TODO: move these tests elsewhere
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::net;
-    use std::time::Duration;
-    use tokio::net::TcpListener;
-
-    #[derive(Debug, Encode, Decode, PartialEq, Eq, Copy, Clone)]
-    struct Transaction {
-        hash: u64,
-        value: u128,
-    }
-
-    #[tokio::test]
-    async fn test_connect_new() {
-        let config = Arc::new(common::chain::config::create_mainnet());
-        let service = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            config,
-            Duration::from_secs(10),
-        )
-        .await;
-        assert!(service.is_ok());
-    }
-
-    // verify that binding to the same interface twice is not possible
-    #[ignore]
-    #[tokio::test]
-    async fn test_connect_new_addrinuse() {
-        let config = Arc::new(common::chain::config::create_mainnet());
-        let service = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            Arc::clone(&config),
-            Duration::from_secs(10),
-        )
-        .await;
-        assert!(service.is_ok());
-
-        let service = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            config,
-            Duration::from_secs(10),
-        )
-        .await;
-
-        match service {
-            Err(e) => {
-                assert_eq!(
-                    e,
-                    P2pError::DialError(DialError::IoError(std::io::ErrorKind::AddrInUse))
-                );
-            }
-            Ok(_) => panic!("address is not in use"),
-        }
-    }
-
-    // try to connect two nodes together by having `service1` listen for network events
-    // and having `service2` trying to connect to `service1`
-    #[tokio::test]
-    async fn test_connect_accept() {
-        let config = Arc::new(common::chain::config::create_mainnet());
-        let service1 = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            Arc::clone(&config),
-            Duration::from_secs(10),
-        )
-        .await;
-        let service2 = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            Arc::clone(&config),
-            Duration::from_secs(10),
-        )
-        .await;
-        assert!(service1.is_ok());
-        assert!(service2.is_ok());
-
-        let (mut service1, _, _) = service1.unwrap();
-        let (mut service2, _, _) = service2.unwrap();
-        let conn_addr = service1.local_addr().clone();
-
-        let (res1, res2): (crate::Result<ConnectivityEvent<Libp2pService>>, _) =
-            tokio::join!(service1.poll_next(), service2.connect(conn_addr));
-
-        assert!(res2.is_ok());
-        assert!(res1.is_ok());
-    }
-
-    // try to connect to a remote peer with a multiaddress that's missing the peerid
-    // and verify that the connection fails
-    #[tokio::test]
-    async fn test_connect_peer_id_missing() {
-        let config = Arc::new(common::chain::config::create_mainnet());
-        let addr: Multiaddr = "/ip6/::1/tcp/8904".parse().unwrap();
-        let (mut service, _, _) = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            config,
-            Duration::from_secs(10),
-        )
-        .await
-        .unwrap();
-
-        match service.connect(addr.clone()).await {
-            Ok(_) => panic!("connect succeeded without peer id"),
-            Err(e) => {
-                assert_eq!(
-                    e,
-                    P2pError::ConversionError(ConversionError::InvalidAddress(addr.to_string()))
-                )
-            }
-        }
-    }
-
-    #[test]
-    fn test_parse_discovered_addr() {
-        let peer_id: PeerId =
-            "12D3KooWE3kBRAnn6jxZMdK1JMWx1iHtR1NKzXSRv5HLTmfD9u9c".parse().unwrap();
-
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/ip4/127.0.0.1/udp/9090/quic".parse().unwrap()),
-            None
-        );
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/ip6/::1/udp/3217".parse().unwrap()),
-            None
-        );
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/ip4/127.0.0.1/tcp/9090/quic".parse().unwrap()),
-            None
-        );
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/ip4/127.0.0.1/tcp/80/http".parse().unwrap()),
-            None
-        );
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/dns4/foo.com/tcp/80/http".parse().unwrap()),
-            None
-        );
-        assert_eq!(
-            parse_discovered_addr(peer_id, "/dns6/foo.com/tcp/443/https".parse().unwrap()),
-            None
-        );
-
-        let addr: Multiaddr =
-            "/ip6/::1/tcp/3217/p2p/12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ"
-                .parse()
-                .unwrap();
-        let id: PeerId = "12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ".parse().unwrap();
-        assert_eq!(parse_discovered_addr(id, addr.clone()), Some(addr));
-
-        let id: PeerId = "12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ".parse().unwrap();
-        let addr: Multiaddr =
-            "/ip4/127.0.0.1/tcp/9090/p2p/12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ"
-                .parse()
-                .unwrap();
-        assert_eq!(parse_discovered_addr(id, addr.clone()), Some(addr));
-
-        let id: PeerId = "12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ".parse().unwrap();
-        let addr: Multiaddr = "/ip6/::1/tcp/3217".parse().unwrap();
-        assert_eq!(
-            parse_discovered_addr(id, addr.clone()),
-            Some(addr.with(Protocol::P2p(id.into())))
-        );
-
-        let id: PeerId = "12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ".parse().unwrap();
-        let addr: Multiaddr = "/ip4/127.0.0.1/tcp/9090".parse().unwrap();
-        assert_eq!(
-            parse_discovered_addr(id, addr.clone()),
-            Some(addr.with(Protocol::P2p(id.into())))
-        );
-    }
-
-    impl PartialEq for Libp2pService {
-        fn eq(&self, _: &Self) -> bool {
-            true
-        }
-    }
-
-    impl<T: NetworkingService> PartialEq for net::types::PeerInfo<T> {
-        fn eq(&self, other: &Self) -> bool {
-            self.peer_id == other.peer_id
-                && self.magic_bytes == other.magic_bytes
-                && self.version == other.version
-                && self.agent == other.agent
-                && self.protocols == other.protocols
-        }
-    }
-
-    // verify that vector of address (that all belong to one peer) parse into one `net::types::Peer` entry
-    #[test]
-    fn test_parse_peers_valid_1_peer() {
-        let id = PeerId::random();
-        let ip4: Multiaddr = "/ip4/127.0.0.1/tcp/9090".parse().unwrap();
-        let ip6: Multiaddr = "/ip6/::1/tcp/9091".parse().unwrap();
-        let addrs = vec![(id, ip4.clone()), (id, ip6.clone())];
-
-        let parsed: Vec<net::types::AddrInfo<Libp2pService>> = parse_peers(addrs);
-        assert_eq!(
-            parsed,
-            vec![net::types::AddrInfo {
-                id,
-                ip4: vec![ip4.with(Protocol::P2p(id.into()))],
-                ip6: vec![ip6.with(Protocol::P2p(id.into()))],
-            }]
-        );
-    }
-
-    // discovery 5 different addresses, ipv4 and ipv6 for both peer and an additional
-    // dns address for peer
-    //
-    // verify that `parse_peers` returns two peers and both only have ipv4 and ipv6 addresses
-    #[test]
-    fn test_parse_peers_valid_2_peers() {
-        let id_1: PeerId = "12D3KooWRn14SemPVxwzdQNg8e8Trythiww1FWrNfPbukYBmZEbJ".parse().unwrap();
-        let ip4_1: Multiaddr = "/ip4/127.0.0.1/tcp/9090".parse().unwrap();
-        let ip6_1: Multiaddr = "/ip6/::1/tcp/9091".parse().unwrap();
-
-        let id_2: PeerId = "12D3KooWE3kBRAnn6jxZMdK1JMWx1iHtR1NKzXSRv5HLTmfD9u9c".parse().unwrap();
-        let ip4_2: Multiaddr = "/ip4/127.0.0.1/tcp/8080".parse().unwrap();
-        let ip6_2: Multiaddr = "/ip6/::1/tcp/8081".parse().unwrap();
-        let dns: Multiaddr = "/dns4/foo.com/tcp/80/http".parse().unwrap();
-
-        let addrs = vec![
-            (id_1, ip4_1.clone()),
-            (id_2, ip4_2.clone()),
-            (id_2, ip6_2.clone()),
-            (id_1, ip6_1.clone()),
-            (id_2, dns),
-        ];
-
-        let mut parsed: Vec<net::types::AddrInfo<Libp2pService>> = parse_peers(addrs);
-        parsed.sort_by(|a, b| a.id.cmp(&b.id));
-
-        assert_eq!(
-            parsed,
-            vec![
-                net::types::AddrInfo {
-                    id: id_2,
-                    ip4: vec![ip4_2.with(Protocol::P2p(id_2.into()))],
-                    ip6: vec![ip6_2.with(Protocol::P2p(id_2.into()))],
-                },
-                net::types::AddrInfo {
-                    id: id_1,
-                    ip4: vec![ip4_1.with(Protocol::P2p(id_1.into()))],
-                    ip6: vec![ip6_1.with(Protocol::P2p(id_1.into()))],
-                },
-            ]
-        );
-    }
-
-    // find 3 peers but only one of the peers have an accepted address available so verify
-    // that `parse_peers()` returns only that peer
-    #[test]
-    fn test_parse_peers_valid_3_peers_1_valid() {
-        let id_1 = PeerId::random();
-        let ip4: Multiaddr = "/ip4/127.0.0.1/tcp/9090".parse().unwrap();
-
-        let id_2 = PeerId::random();
-        let dns: Multiaddr = "/dns4/foo.com/tcp/80/http".parse().unwrap();
-
-        let id_3 = PeerId::random();
-        let quic: Multiaddr = "/ip4/127.0.0.1/tcp/9090/quic".parse().unwrap();
-
-        let addrs = vec![(id_1, ip4.clone()), (id_2, dns), (id_3, quic)];
-        let parsed: Vec<net::types::AddrInfo<Libp2pService>> = parse_peers(addrs);
-
-        assert_eq!(
-            parsed,
-            vec![net::types::AddrInfo {
-                id: id_1,
-                ip4: vec![ip4.with(Protocol::P2p(id_1.into()))],
-                ip6: vec![],
-            }]
-        );
-    }
-
-    // try to connect to a service that is not listening with a small timeout and verify that the connection fails
-    #[tokio::test]
-    async fn test_connect_with_timeout() {
-        let config = Arc::new(common::chain::config::create_mainnet());
-        let (mut service, _, _) = Libp2pService::start(
-            test_utils::make_address("/ip6/::1/tcp/"),
-            &[],
-            &[],
-            config,
-            Duration::from_secs(2),
-        )
-        .await
-        .unwrap();
-
-        let port = portpicker::pick_unused_port().unwrap();
-        let mut addr: Multiaddr = format!("/ip6/::1/tcp/{}", port).parse().unwrap();
-        addr.push(Protocol::P2p(PeerId::random().into()));
-
-        // first try to connect to address nobody is listening to
-        // and verify that the connection is refused immediately
-        let start = std::time::SystemTime::now();
-        assert_eq!(service.connect(addr.clone()).await, Ok(()));
-        assert!(std::matches!(
-            service.poll_next().await,
-            Ok(net::types::ConnectivityEvent::ConnectionError {
-                addr: _,
-                error: P2pError::DialError(DialError::IoError(
-                    std::io::ErrorKind::ConnectionRefused
-                ))
-            })
-        ));
-
-        let timeout = if cfg!(target_os = "linux") || cfg!(target_os = "macos") {
-            0
-        } else {
-            2
-        };
-        assert_eq!(
-            std::time::SystemTime::now().duration_since(start).unwrap().as_secs(),
-            timeout
-        );
-
-        // then create a socket that listens to the address and verify that it takes
-        // 2 seconds to get the `ConnectionRefused` error, as expected
-        let _service = TcpListener::bind(format!("[::1]:{}", port)).await.unwrap();
-        let start = std::time::SystemTime::now();
-
-        assert_eq!(service.connect(addr).await, Ok(()),);
-        assert!(std::matches!(
-            service.poll_next().await,
-            Ok(net::types::ConnectivityEvent::ConnectionError {
-                addr: _,
-                error: P2pError::DialError(DialError::IoError(
-                    std::io::ErrorKind::ConnectionRefused
-                ))
-            })
-        ));
-        assert!(std::time::SystemTime::now().duration_since(start).unwrap().as_secs() >= 2);
     }
 }
