--- p2p/src/net/libp2p/proto/mod.rs
@@ -1,23 +1,0 @@
-// Copyright (c) 2021 Protocol Labs
-// Copyright (c) 2022 RBB S.r.l
-// opensource@mintlayer.org
-// SPDX-License-Identifier: MIT
-// Licensed under the MIT License;
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://spdx.org/licenses/MIT
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Author(s): A. Altonen
-pub mod connectivity;
-pub mod gossipsub;
-pub mod identify;
-pub mod mdns;
-pub mod ping;
-pub mod util;
